move find_isosurface to analysis subpackage
