+++ src/cdiutils/analysis/__init__.py
@@ -1,8 +1,9 @@
-from .stats import get_histogram, plot_histogram
+from .stats import get_histogram, plot_histogram, find_isosurface
 from .stereography import pole_figure
 
 __all__ = [
     "get_histogram",
     "plot_histogram",
+    "find_isosurface",
     "pole_figure"
 ]
