+++ src/cdiutils/utils.py
@@ -4,12 +4,8 @@ import warnings
 import numpy as np
 from scipy.fft import fftshift, ifftshift, fft2, ifft2, fftn
 import matplotlib
-import seaborn as sns
 import scipy.constants as cts
 from scipy.ndimage import convolve, center_of_mass, median_filter
-from scipy.stats import gaussian_kde
-
-from cdiutils.plot.formatting import save_fig
 
 
 def bin_along_axis(
@@ -1102,115 +1098,6 @@ def extract_reduced_shape(
     return shape
 
 
-def find_isosurface(
-        amplitude: np.ndarray,
-        nbins: int = 100,
-        sigma_criterion: float = 3,
-        plot: bool = False,
-        show: bool = False,
-        save: str = None
-) -> tuple[float, matplotlib.axes.Axes] | float:
-    """
-    Estimate the isosurface from the amplitude distribution
-
-    :param amplitude: the 3D amplitude volume (np.array)
-    :param nbins: the number of bins to considerate when making the
-    histogram (optional, int)
-    :param sigma_criterion: the factor to compute the isosurface which is
-    calculated as: mu - sigma_criterion * sigma. By default set to 3.
-    (optional, float)
-    :param show: whether or not to show the the figure
-    :param save: where to save the figure is plotted
-
-    :return: the isosurface value and the figure in which the histogram
-    was plotted
-    """
-    # normalise and flatten the amplitude
-    flattened_amplitude = normalise(amplitude).ravel()
-
-    counts, bins = np.histogram(flattened_amplitude, bins=nbins)
-
-    # remove the background
-    background_value = bins[
-        np.where(counts == counts.max())[0] + 1 + nbins//20
-    ]
-    filtered_amplitude = flattened_amplitude[
-        flattened_amplitude > background_value
-    ]
-
-    # redo the histogram with the filtered amplitude
-    counts, bins = np.histogram(filtered_amplitude, bins=nbins, density=True)
-    bin_centres = (bins[:-1] + bins[1:]) / 2
-    bin_size = bin_centres[1] - bin_centres[0]
-
-    # fit the amplitude distribution
-    kernel = gaussian_kde(filtered_amplitude)
-    x = np.linspace(0, 1, 1000)
-    fitted_counts = kernel(x)
-
-    max_index = np.argmax(fitted_counts)
-    right_gaussian_part = np.where(x >= x[max_index], fitted_counts, 0)
-
-    # find the closest indexes
-    right_HM_index = np.argmin(
-        np.abs(right_gaussian_part - fitted_counts.max() / 2)
-    )
-    left_HM_index = max_index - (right_HM_index - max_index)
-
-    fwhm = x[right_HM_index] - x[left_HM_index]
-    sigma_estimate = fwhm / 2*np.sqrt(2*np.log(2))
-    isosurface = x[max_index] - sigma_criterion * sigma_estimate
-
-    if plot or show:
-        figsize = (6, 4)  # (5.812, 3.592)  # golden ratio
-        fig, ax = matplotlib.pyplot.subplots(1, 1, layout="tight", figsize=figsize)
-        ax.bar(
-            bin_centres,
-            counts,
-            width=bin_size,
-            color="dodgerblue",
-            alpha=0.9,
-            edgecolor=(0, 0, 0, 0.25),
-            label=r"amplitude distribution"
-        )
-        sns.kdeplot(
-            filtered_amplitude,
-            ax=ax,
-            alpha=0.3,
-            fill=True,
-            color="navy",
-            label=r"density estimate"
-        )
-        ax.axvspan(
-            x[left_HM_index],
-            x[right_HM_index],
-            edgecolor="k",
-            facecolor="green",
-            alpha=0.2,
-            label="FWHM"
-        )
-        ax.plot(
-            [isosurface, isosurface],
-            [0, fitted_counts[(np.abs(x - isosurface)).argmin()]],
-            solid_capstyle="round",
-            color="lightcoral",
-            lw=5,
-            label=fr"isosurface estimated at {isosurface:0.3f}"
-        )
-
-        ax.set_xlabel(r"normalised amplitude")
-        ax.set_ylabel("counts")
-        ax.legend(frameon=False)
-        fig.suptitle(r"Reconstructed amplitude distribution")
-        fig.tight_layout()
-        if save is not None:
-            save_fig(fig, save, transparent=False)
-        if show:
-            matplotlib.pyplot.show()
-        return float(isosurface), fig
-    return float(isosurface)
-
-
 def get_oversampling_ratios(
     support: np.ndarray = None,
     direct_space_object: np.ndarray = None,
@@ -1359,65 +1246,6 @@ def hot_pixel_filter(
     mask = data < threshold * (data_median + 1)
     cleaned_data = data * mask
     return cleaned_data, np.logical_not(mask)
-
-
-def kde_from_histogram(
-        counts: np.ndarray,
-        bin_edges: np.ndarray,
-) -> tuple[np.ndarray, np.ndarray]:
-    """
-    Compute the Kernel Density Estimate (KDE) from histogram counts and
-    bin edges provided by numpy.histogram function.
-
-    Args:
-        counts (np.ndarray): the number of elements in each bin.
-        bin_edges (np.ndarray): the limits of each bin.
-
-    Returns:
-        tuple[np.ndarray, np.ndarray]: x values used to compute the KDE
-            estimate, the y value (KDE estimate)
-    """
-    # Check if the histogram is density or not by checking the sum of
-    # the counts
-    bin_widths = np.diff(bin_edges)
-    is_density = np.isclose(np.sum(counts * bin_widths), 1.0)
-
-    if is_density:
-        # When density=True, use the bin edges to reconstruct the data
-        # for KDE
-        data = []
-        for count, left_edge, right_edge in zip(
-                counts, bin_edges[:-1], bin_edges[1:]
-        ):
-            data.extend(
-                np.random.uniform(
-                    left_edge,
-                    right_edge,
-                    int(count * len(counts) * (right_edge - left_edge))
-                )
-            )
-        data = np.array(data)
-
-        kde = gaussian_kde(data)
-        x = np.linspace(min(bin_edges), max(bin_edges))
-        y = kde(x)
-
-    else:
-        # Reconstruct the data from histogram counts and bin edges
-        bin_centers = (bin_edges[:-1] + bin_edges[1:]) / 2
-        bin_width = bin_edges[1] - bin_edges[0]
-        reconstructed_data = np.repeat(bin_centers, counts)
-
-        # Calculate KDE using the reconstructed data
-        kde = gaussian_kde(reconstructed_data)
-        # Evaluate KDE
-        x = np.linspace(bin_edges.min(), bin_edges.max())
-        y = kde.pdf(x)
-
-        # Scale the KDE values to match the original counts
-        y *= len(reconstructed_data) * bin_width
-
-    return x, y
 
 
 def valid_args_only(params: dict, function: callable) -> dict:
